refactor: reduce complexity of task 2 file by moving methods back to core
